feat: Game엔티티 클래스에 create, update 메서드 추가

+++ backend/src/main/kotlin/com/backend/infra/steam/dto/response/SteamAppListResponse.kt
@@ -3,7 +3,7 @@ package com.backend.infra.steam.dto.response
 import com.fasterxml.jackson.annotation.JsonProperty
 
 data class SteamApp(
-    val appid: Int,
+    val appid: Long,
     val name: String,
     @JsonProperty("last_modified")
     val lastModified: Long,
